Added support for alembic meshes (www.alembic.io)
Alembic is an open source file format for exchanging 3D computer animation information between different platforms and applications. Any app can write their 3D information out in the Alembic format which can then be read in by any other app that supports Alembic, such as Maya, 3dsmax, Houdini, Katana, Nuke to name a few. Beyond that, it stores 3D geometry in a format that is very efficient with a small file size that is quick to read and unpack.